+++ app/client/cypress/locators/ApiEditor.js
@@ -1,12 +1,11 @@
 export default {
-  curlImage: ".t--curlImage",
   curlImportBtn: ".t--importBtn",
   createBlankApiCard: ".t--createBlankApiCard",
   eachProviderCard: ".t--eachProviderCard",
   nameOfApi: ".t--nameOfApi",
   ApiNameField: ".t--action-name-edit-field",
   addToPageBtn: ".t--addToPageBtn",
-  ApiActionMenu: '[data-testid="more-action-trigger"]',
+  ApiActionMenu: "[data-testid=\"more-action-trigger\"]",
   ApiDeleteBtn: ".t--apiFormDeleteBtn",
   ApiRunBtn: ".t--apiFormRunBtn",
   addToPageBtnsId: ".t--addToPageButtons",
@@ -23,7 +22,7 @@ export default {
     ".t--apiFormPaginationType label:contains('Paginate with response URL') input",
   apiTab: ".react-tabs__tab-list li",
   bodyType: ".t--apiFormPostBodyType",
-  bodyTypeSelected: '[data-testid="t--api-body-tab-switch"] .rc-select-selection-item',
+  bodyTypeSelected: "[data-testid=\"t--api-body-tab-switch\"] .rc-select-selection-item",
   bodyTab: "Body",
   headersTab: "Header",
   jsonResponseTab: "[data-value='JSON']",
@@ -34,5 +33,5 @@ export default {
   apiSearchHint: ".datasource-hint",
   slashCommandButton: ".commands-button",
   apiResponseObject: ".object-key",
-  apiDebuggerLink: ".debugger-entity-link"
+  apiDebuggerLink: ".debugger-entity-link",
 };
