+++ app/client/src/ce/entities/IDE/constants.ts
@@ -5,7 +5,6 @@ import {
   BUILDER_CUSTOM_PATH,
   BUILDER_PATH,
   BUILDER_PATH_DEPRECATED,
-  CURL_IMPORT_PAGE_PATH,
   DATA_SOURCES_EDITOR_ID_PATH,
   ENTITY_PATH,
   INTEGRATION_EDITOR_PATH,
@@ -127,8 +126,6 @@ export const EntityPaths: string[] = [
   JS_COLLECTION_ID_ADD_PATH,
   WIDGETS_EDITOR_ID_PATH,
   WIDGETS_EDITOR_ID_PATH + ADD_PATH,
-  CURL_IMPORT_PAGE_PATH,
-  CURL_IMPORT_PAGE_PATH + ADD_PATH,
   ENTITY_PATH,
 ];
