+++ app/client/src/ce/pages/Editor/IDE/EditorPane/constants.ts
@@ -1,6 +1,5 @@
 import {
   API_EDITOR_BASE_PATH,
-  CURL_IMPORT_PAGE_PATH,
   JS_COLLECTION_EDITOR_PATH,
   QUERIES_EDITOR_BASE_PATH,
   WIDGETS_EDITOR_BASE_PATH,
@@ -8,7 +7,6 @@ import {
 import { SAAS_EDITOR_PATH } from "pages/Editor/SaaSEditor/constants";
 
 export const querySegmentRoutes = [
-  CURL_IMPORT_PAGE_PATH,
   API_EDITOR_BASE_PATH,
   `${SAAS_EDITOR_PATH}/api`,
   QUERIES_EDITOR_BASE_PATH,
