+++ app/client/src/pages/Editor/APIEditor/helpers.ts
@@ -1,20 +1,3 @@
-import { submitCurlImportForm } from "actions/importActions";
-import type { ActionParentEntityTypeInterface } from "@appsmith/entities/Engine/actionHelpers";
-
-export interface curlImportFormValues {
-  curl: string;
-  contextId: string;
-  name: string;
-  contextType: ActionParentEntityTypeInterface;
-}
-
-export const curlImportSubmitHandler = (
-  values: curlImportFormValues,
-  dispatch: any,
-) => {
-  dispatch(submitCurlImportForm(values));
-};
-
 export const sortedDatasourcesHandler = (
   datasources: Record<string, any>,
   currentDatasourceId: string,
