+++ app/client/src/reducers/uiReducers/importReducer.ts
@@ -6,11 +6,20 @@ import {
 } from "@appsmith/constants/ReduxActionConstants";
 
 const initialState: ImportReduxState = {
+  isCurlModalOpen: false,
   isImportingCurl: false,
   errorPayload: {},
 };
 
 const importReducer = createReducer(initialState, {
+  [ReduxActionTypes.SET_CURL_MODAL_OPEN]: (state: ImportReduxState) => ({
+    ...state,
+    isCurlModalOpen: true,
+  }),
+  [ReduxActionTypes.SET_CURL_MODAL_CLOSE]: (state: ImportReduxState) => ({
+    ...state,
+    isCurlModalOpen: false,
+  }),
   [ReduxActionTypes.SUBMIT_CURL_FORM_INIT]: (state: ImportReduxState) => {
     return {
       ...state,
@@ -20,6 +29,7 @@ const importReducer = createReducer(initialState, {
   [ReduxActionTypes.SUBMIT_CURL_FORM_SUCCESS]: (state: ImportReduxState) => ({
     ...state,
     isImportingCurl: false,
+    isCurlModalOpen: false,
   }),
   [ReduxActionErrorTypes.SUBMIT_CURL_FORM_ERROR]: (
     state: ImportReduxState,
@@ -30,6 +40,7 @@ const importReducer = createReducer(initialState, {
 });
 
 export interface ImportReduxState {
+  isCurlModalOpen: boolean;
   isImportingCurl: boolean;
   errorPayload: Record<string, unknown>;
 }
