+++ app/client/src/selectors/ui.tsx
@@ -63,9 +63,6 @@ export const getDatasourceCollapsibleState = createSelector(
   },
 );
 
-export const getIsImportingCurl = (state: AppState) =>
-  state.ui.imports.isImportingCurl;
-
 export const getIsConsolidatedPageLoading = (state: AppState) =>
   state.ui.consolidatedPageLoad.isLoading;
